Move internal functions closer to their point of use
