+++ wandb/sdk/internal/artifacts.py
@@ -1,27 +1,23 @@
 import concurrent.futures
 import json
 import os
-import sys
 import tempfile
 from typing import TYPE_CHECKING, Any, Dict, List, Optional, Sequence
 
 import wandb
 import wandb.filesync.step_prepare
-from wandb import util
+from wandb import env, util
+from wandb.sdk.interface.artifacts import ArtifactManifest, ArtifactManifestEntry
+from wandb.sdk.lib.filesystem import mkdir_exists_ok
 from wandb.sdk.lib.hashutil import B64MD5, b64_to_hex_id, md5_file_b64
-
-from ..interface.artifacts import (
-    ArtifactManifest,
-    ArtifactManifestEntry,
-    get_staging_dir,
-)
+from wandb.util import FilePathStr, URIStr
 
 if TYPE_CHECKING:
-    from wandb.proto import wandb_internal_pb2
+    import sys
+
+    from wandb.sdk.internal.file_pusher import FilePusher
     from wandb.sdk.internal.internal_api import Api as InternalApi
     from wandb.sdk.internal.progress import ProgressFn
-
-    from .file_pusher import FilePusher
 
     if sys.version_info >= (3, 8):
         from typing import Protocol
@@ -35,35 +31,10 @@ if TYPE_CHECKING:
             pass
 
 
-def _manifest_json_from_proto(manifest: "wandb_internal_pb2.ArtifactManifest") -> Dict:
-    if manifest.version == 1:
-        contents = {
-            content.path: {
-                "digest": content.digest,
-                "birthArtifactID": content.birth_artifact_id
-                if content.birth_artifact_id
-                else None,
-                "ref": content.ref if content.ref else None,
-                "size": content.size if content.size is not None else None,
-                "local_path": content.local_path if content.local_path else None,
-                "extra": {
-                    extra.key: json.loads(extra.value_json) for extra in content.extra
-                },
-            }
-            for content in manifest.contents
-        }
-    else:
-        raise Exception(f"unknown artifact manifest version: {manifest.version}")
-
-    return {
-        "version": manifest.version,
-        "storagePolicy": manifest.storage_policy,
-        "storagePolicyConfig": {
-            config.key: json.loads(config.value_json)
-            for config in manifest.storage_policy_config
-        },
-        "contents": contents,
-    }
+def get_staging_dir() -> FilePathStr:
+    path = os.path.join(env.get_data_dir(), "artifacts", "staging")
+    mkdir_exists_ok(path)
+    return FilePathStr(os.path.abspath(os.path.expanduser(path)))
 
 
 class ArtifactSaver:
@@ -302,7 +273,7 @@ class ArtifactSaver:
                     artifact_id = self._api._resolve_client_id(client_id)
                     if artifact_id is None:
                         raise RuntimeError(f"Could not resolve client id {client_id}")
-                    entry.ref = util.URIStr(
+                    entry.ref = URIStr(
                         "wandb-artifact://{}/{}".format(
                             b64_to_hex_id(B64MD5(artifact_id)), artifact_file_path
                         )
